+++ lib/tuple.dart
@@ -468,7 +468,7 @@ class Tuple7<T1, T2, T3, T4, T5, T6, T7> {
       ]);
 }
 
-/// Represents a 8-tuple, or septuple.
+/// Represents a 8-tuple, or octuple.
 class Tuple8<T1, T2, T3, T4, T5, T6, T7, T8> {
   /// Returns the first item of the tuple
   final T1 item1;
@@ -589,5 +589,141 @@ class Tuple8<T1, T2, T3, T4, T5, T6, T7, T8> {
         item6.hashCode,
         item7.hashCode,
         item8.hashCode
+      ]);
+}
+
+/// Represents a 9-tuple, or nonuple.
+class Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> {
+  /// Returns the first item of the tuple
+  final T1 item1;
+
+  /// Returns the second item of the tuple
+  final T2 item2;
+
+  /// Returns the third item of the tuple
+  final T3 item3;
+
+  /// Returns the fourth item of the tuple
+  final T4 item4;
+
+  /// Returns the fifth item of the tuple
+  final T5 item5;
+
+  /// Returns the sixth item of the tuple
+  final T6 item6;
+
+  /// Returns the seventh item of the tuple
+  final T7 item7;
+
+  /// Returns the eight item of the tuple
+  final T8 item8;
+
+  /// Returns the ninth item of the tuple
+  final T9 item9;
+
+  /// Creates a new tuple value with the specified items.
+  const Tuple9(this.item1, this.item2, this.item3, this.item4, this.item5,
+      this.item6, this.item7, this.item8, this.item9);
+
+  /// Create a new tuple value with the specified list [items].
+  factory Tuple9.fromList(List items) {
+    if (items.length != 9) {
+      throw ArgumentError('items must have length 9');
+    }
+
+    return Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+        items[0] as T1,
+        items[1] as T2,
+        items[2] as T3,
+        items[3] as T4,
+        items[4] as T5,
+        items[5] as T6,
+        items[6] as T7,
+        items[7] as T8,
+        items[8] as T9);
+  }
+
+  /// Returns a tuple with the first item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem1(T1 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          v, item2, item3, item4, item5, item6, item7, item8, item9);
+
+  /// Returns a tuple with the second item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem2(T2 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, v, item3, item4, item5, item6, item7, item8, item9);
+
+  /// Returns a tuple with the third item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem3(T3 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, v, item4, item5, item6, item7, item8, item9);
+
+  /// Returns a tuple with the fourth item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem4(T4 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, item3, v, item5, item6, item7, item8, item9);
+
+  /// Returns a tuple with the fifth item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem5(T5 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, item3, item4, v, item6, item7, item8, item9);
+
+  /// Returns a tuple with the sixth item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem6(T6 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, item3, item4, item5, v, item7, item8, item9);
+
+  /// Returns a tuple with the seventh item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem7(T7 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, item3, item4, item5, item6, v, item8, item9);
+
+  /// Returns a tuple with the eight item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem8(T8 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, item3, item4, item5, item6, item7, v, item9);
+
+  /// Returns a tuple with the ninth item set to the specified value.
+  Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9> withItem9(T9 v) =>
+      Tuple9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
+          item1, item2, item3, item4, item5, item6, item7, item8, v);
+
+  /// Creates a [List] containing the items of this [Tuple9].
+  ///
+  /// The elements are in item order. The list is variable-length
+  /// if [growable] is true.
+  List toList({bool growable = false}) =>
+      List.from([item1, item2, item3, item4, item5, item6, item7, item8, item9],
+          growable: growable);
+
+  @override
+  String toString() =>
+      // ignore: lines_longer_than_80_chars
+      '[$item1, $item2, $item3, $item4, $item5, $item6, $item7, $item8, $item9]';
+
+  @override
+  bool operator ==(Object other) =>
+      other is Tuple9 &&
+      other.item1 == item1 &&
+      other.item2 == item2 &&
+      other.item3 == item3 &&
+      other.item4 == item4 &&
+      other.item5 == item5 &&
+      other.item6 == item6 &&
+      other.item7 == item7 &&
+      other.item8 == item8 &&
+      other.item9 == item9;
+
+  @override
+  int get hashCode => hashObjects([
+        item1.hashCode,
+        item2.hashCode,
+        item3.hashCode,
+        item4.hashCode,
+        item5.hashCode,
+        item6.hashCode,
+        item7.hashCode,
+        item8.hashCode,
+        item9.hashCode,
       ]);
 }